Add gloabl search for smeagol documents (#11)
This will index markdown documents in the configured smeagol document folder for repositories with a valid .smeagol.yml configuration file. This can be used by smeagol to run a "global" search over all documents without having to check out all repositories.

In this first implementation, only the default branch of a repository will be taken into account. Other branches shall be handled in the future.

Co-authored-by: Matthias Thieroff <matthias.thieroff@cloudogu.com>